Upgrade example to latest standards
