+++ example/pubspec.yaml
@@ -8,21 +8,21 @@ environment:
   sdk: ">=2.17.0 <4.0.0"
 
 dependencies:
-  shared_preferences: ^2.0.13
+  shared_preferences: ^2.2.2
   flutter_colorpicker: ^1.0.3
-  url_launcher: ^6.1.0
+  url_launcher: ^6.2.4
   auto_size_text: ^3.0.0
   flutter:
     sdk: flutter
   dashboard:
     path: ..
-  cupertino_icons: ^1.0.4
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^1.0.4
+  flutter_lints: ^3.0.1
 
 flutter:
   uses-material-design: true
